nuevos cambios en app ventas - flutter

+++ lib/data/repository/usuario_provider.dart
@@ -1,40 +1,82 @@
 import 'package:flutter/material.dart';
 
+import 'package:app_ventas/helper/apiRest.dart';
+import 'package:app_ventas/data/dao/login_dao.dart';
+import 'package:app_ventas/data/dao/cliente_dao.dart';
+
+import 'package:app_ventas/data/models/sync_model.dart';
+import 'package:app_ventas/data/models/login_model.dart';
+import 'package:app_ventas/data/shared_prefs/preferencias_usuario.dart';
+
 import 'dart:convert';
 
-import 'package:app_amsys/src/dao/login_dao.dart';
-import 'package:app_amsys/src/shared_prefs/preferencias_usuario.dart';
 import 'package:http/http.dart' as http;
-
-import 'package:app_amsys/src/models/login_model.dart';
+//import 'package:package_info/package_info.dart';
 
 class UsuarioProvider with ChangeNotifier {
-  final String _url = 'http://alphaman-001-site8.ftempurl.com/api';
   final _prefs = new PreferenciasUsuario();
 
   Future<Map<String, dynamic>> login(String email, String password) async {
-    final authData = {'User': email, 'Password': password};
+    //final info = await PackageInfo.fromPlatform();
 
-    final resp = await http.post('$_url/usuario/login',
+    final authData = {
+      "login": email,
+      "pass": password,
+      "imei": "123",
+      "version": "2.7"
+    };
+
+    final resp = await http.post('$url/LoginNew',
         body: json.encode(authData),
         headers: <String, String>{
           'Content-Type': 'application/json; charset=UTF-8',
         });
 
-    final loginResponse = loginModelFromJson(resp.body);
-
-    if (loginResponse.response.codigo == "0000") {
-      await LoginDao().nuevoUsuario(loginResponse.data);
-      _prefs.login = 'home';
-      return {'ok': true};
-    } else {
+    if (resp.statusCode != 200 && resp.statusCode != 201) {
       _prefs.login = 'login';
-      return {'ok': false, 'mensaje': loginResponse.response.comentario};
+      return {'ok': false, 'mensaje': resp.body};
     }
+
+    final loginResponse = usuarioFromJson(resp.body);
+    await LoginDao().nuevoUsuario(loginResponse);
+    _prefs.login = 'home';
+    return {'ok': true};
   }
 
-  Future<void> logout() async {
+  Future<Map<String, dynamic>> logout() async {
+    final u = await LoginDao().getUsuario();
+    final authData = {"login": u.login};
+
+    final resp = await http.post('$url/Logout',
+        body: json.encode(authData),
+        headers: <String, String>{
+          'Content-Type': 'application/json; charset=UTF-8',
+        });
+
+    if (resp.statusCode != 200 && resp.statusCode != 201) {
+      return {'ok': false, 'mensaje': resp.body};
+    }
+
     await LoginDao().deleteAllUsuario();
+    await ClienteDao().deleteAllCliente();
     _prefs.login = 'login';
+    return {'ok': true};
+  }
+
+  Future<Map<String, dynamic>> sincronizacion() async {
+    final u = await LoginDao().getUsuario();
+
+    final resp =
+        await http.get('$url/Sync?operarioId=${u.usuarioId}&version=2.7');
+
+    if (resp.statusCode != 200 && resp.statusCode != 201) {
+      _prefs.login = 'login';
+      return {'ok': false, 'mensaje': resp.body};
+    }
+
+    final syncResponse = syncFromJson(resp.body);
+    await ClienteDao().insertClientes(syncResponse.clientes);
+    _prefs.login = 'home';
+    return {'ok': true};
   }
 }
